Add superscript '?' links to a few productions
Note: Just adding 4 references for now until we get further into grammar
refactoring.

Now that the productions have moved to Appendix A, it should be easy to
find them from the places they used to be.

This commit adds links like `[#c-printable]` to add a superscript `?`
link to the production supporting the text.

When we refactor the production names, m2k will warn in RED about
orphaned links. That way we can work on grammar refactoring without it
blocking on cross-referencing. Then when the dust has settled the logs
will tell us where fixes need to go.