Combine two comment productions that appear to do same thing
As far as I can tell maybe-comment-and-end-of-line and
blanks-and-maybe-comment-and-end-of-line appear to do the same thing.

Put another way, I can't think of an situation using this production set
where there needs to be both.

I renamed them to comment-line which matches all of:

* (one or more blanks) or start-of-line
* possibly a ('#' and all the following non-break characters)
* (always) a line break

This commit needs careful review. To refute it, find a YAML text that
would cause the productions to not work as before.